Got it working, one problem though, see desc
the values of x, y, z are rounded, so it causes a problem when the users
are teleported back. I tested this out by standing on the 2nd level of
the arena, when I was teleported, i ended up below. This ccould end up
being a problem when players figure this out and end up in a base they
want to raid